requirements_stablediffusion.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
pillow
numpy

